+++ src/main/java/io/github/mzmine/gui/chartbasics/template/SimpleToolTipGenerator.java
@@ -16,16 +16,16 @@
  * USA
  */
 
-package io.github.mzmine.gui.chartbasics.gui.javafx.template;
+package io.github.mzmine.gui.chartbasics.template;
 
 import io.github.mzmine.gui.chartbasics.gui.javafx.EChartViewer;
-import io.github.mzmine.gui.chartbasics.gui.javafx.template.providers.ToolTipTextProvider;
+import io.github.mzmine.gui.chartbasics.template.providers.ToolTipTextProvider;
 import org.jfree.chart.labels.XYToolTipGenerator;
 import org.jfree.chart.plot.XYPlot;
 import org.jfree.data.xy.XYDataset;
 
 /**
- * Default tooltip generator. Generates tooltips based on {@link io.github.mzmine.gui.chartbasics.gui.javafx.template.providers.ToolTipTextProvider#getToolTipText(int)}.
+ * Default tooltip generator. Generates tooltips based on {@link io.github.mzmine.gui.chartbasics.template.providers.ToolTipTextProvider#getToolTipText(int)}.
  *
  * @author https://github.com/SteffenHeu
  */
