+++ src/main/java/io/github/mzmine/gui/chartbasics/template/SimpleXYLineChartWithDatasetView.java
@@ -16,9 +16,9 @@
  * USA
  */
 
-package io.github.mzmine.gui.chartbasics.gui.javafx.template;
+package io.github.mzmine.gui.chartbasics.template;
 
-import io.github.mzmine.gui.chartbasics.gui.javafx.template.providers.PlotXYDatasetProvider;
+import io.github.mzmine.gui.chartbasics.template.providers.PlotXYDatasetProvider;
 import javafx.geometry.Orientation;
 import javafx.scene.control.SplitPane;
 import javax.annotation.Nonnull;
